+++ src/main/java/com/github/theredbrain/healthyfood/HealthyFood.java
@@ -1,17 +1,14 @@
 package com.github.theredbrain.healthyfood;
 
 import com.github.theredbrain.healthyfood.config.ServerConfig;
-import com.github.theredbrain.healthyfood.config.ServerConfigWrapper;
-import com.google.gson.Gson;
-import me.shedaniel.autoconfig.AutoConfig;
-import me.shedaniel.autoconfig.serializer.JanksonConfigSerializer;
-import me.shedaniel.autoconfig.serializer.PartitioningSerializer;
+import com.github.theredbrain.manaattributes.entity.ManaUsingEntity;
+import com.github.theredbrain.staminaattributes.entity.StaminaUsingEntity;
+import me.fzzyhmstrs.fzzy_config.api.ConfigApiJava;
+import me.fzzyhmstrs.fzzy_config.api.RegisterType;
 import net.fabricmc.api.ModInitializer;
-
-import net.fabricmc.fabric.api.networking.v1.PacketByteBufs;
-import net.fabricmc.fabric.api.networking.v1.ServerPlayConnectionEvents;
+import net.fabricmc.loader.api.FabricLoader;
+import net.minecraft.entity.LivingEntity;
 import net.minecraft.entity.attribute.EntityAttribute;
-import net.minecraft.network.PacketByteBuf;
 import net.minecraft.util.Identifier;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -19,44 +16,49 @@ import org.slf4j.LoggerFactory;
 public class HealthyFood implements ModInitializer {
 	public static final String MOD_ID = "healthyfood";
 	public static final Logger LOGGER = LoggerFactory.getLogger(MOD_ID);
-	public static ServerConfig serverConfig;
-	private static PacketByteBuf serverConfigSerialized = PacketByteBufs.create();
+	public static ServerConfig SERVER_CONFIG;
 
 	public static EntityAttribute FULLNESS_TICK_THRESHOLD;
 	public static EntityAttribute MAX_FULLNESS;
+
+	public static final boolean isManaAttributesLoaded = FabricLoader.getInstance().isModLoaded("manaattributes");
+	public static final boolean isStaminaAttributesLoaded = FabricLoader.getInstance().isModLoaded("staminaattributes");
+
+	public static float getCurrentMana(LivingEntity livingEntity) {
+		float currentMana = 0.0F;
+		if (isManaAttributesLoaded) {
+			currentMana = ((ManaUsingEntity) livingEntity).manaattributes$getMana();
+		}
+		return currentMana;
+	}
+
+	public static void addMana(LivingEntity livingEntity, float amount) {
+		if (isManaAttributesLoaded) {
+			((ManaUsingEntity) livingEntity).manaattributes$addMana(amount);
+		}
+	}
+
+	public static float getCurrentStamina(LivingEntity livingEntity) {
+		float currentStamina = 0.0F;
+		if (isStaminaAttributesLoaded) {
+			currentStamina = ((StaminaUsingEntity) livingEntity).staminaattributes$getStamina();
+		}
+		return currentStamina;
+	}
+
+	public static void addStamina(LivingEntity livingEntity, float amount) {
+		if (isStaminaAttributesLoaded) {
+			((StaminaUsingEntity) livingEntity).staminaattributes$addStamina(amount);
+		}
+	}
 
 	@Override
 	public void onInitialize() {
 		LOGGER.info("Eat healthy!");
 
 		// Config
-		AutoConfig.register(ServerConfigWrapper.class, PartitioningSerializer.wrap(JanksonConfigSerializer::new));
-		serverConfig = ((ServerConfigWrapper) AutoConfig.getConfigHolder(ServerConfigWrapper.class).getConfig()).server;
+		SERVER_CONFIG = ConfigApiJava.registerAndLoadConfig(ServerConfig::new, RegisterType.BOTH);
 
-		// Events
-		serverConfigSerialized = ServerConfigSync.write(serverConfig);
-		ServerPlayConnectionEvents.JOIN.register((handler, sender, server) -> {
-			sender.sendPacket(ServerConfigSync.ID, serverConfigSerialized);
-		});
-
-	}
-
-	public static class ServerConfigSync {
-		public static Identifier ID = identifier("server_config_sync");
-
-		public static PacketByteBuf write(ServerConfig serverConfig) {
-			var gson = new Gson();
-			var json = gson.toJson(serverConfig);
-			var buffer = PacketByteBufs.create();
-			buffer.writeString(json);
-			return buffer;
-		}
-
-		public static ServerConfig read(PacketByteBuf buffer) {
-			var gson = new Gson();
-			var json = buffer.readString();
-			return gson.fromJson(json, ServerConfig.class);
-		}
 	}
 
 	public static Identifier identifier(String path) {