+++ src/main/java/com/github/theredbrain/healthyfood/mixin/entity/attribute/EntityAttributesMixin.java
@@ -9,13 +9,13 @@ import org.spongepowered.asm.mixin.Shadow;
 
 @Mixin(EntityAttributes.class)
 public class EntityAttributesMixin {
-    @Shadow
-    private static EntityAttribute register(String id, EntityAttribute attribute) {
-        throw new AssertionError();
-    }
+	@Shadow
+	private static EntityAttribute register(String id, EntityAttribute attribute) {
+		throw new AssertionError();
+	}
 
-    static {
-        HealthyFood.FULLNESS_TICK_THRESHOLD = register(HealthyFood.MOD_ID + ":generic.fullness_tick_threshold", new ClampedEntityAttribute("attribute.name.generic.fullness_tick_threshold", 400.0, 0.0, 1024.0).setTracked(true));
-        HealthyFood.MAX_FULLNESS = register(HealthyFood.MOD_ID + ":generic.max_fullness", new ClampedEntityAttribute("attribute.name.generic.max_fullness", 3.0, 0.0, 1024.0).setTracked(true));
-    }
+	static {
+		HealthyFood.FULLNESS_TICK_THRESHOLD = register(HealthyFood.MOD_ID + ":generic.fullness_tick_threshold", new ClampedEntityAttribute("attribute.name.generic.fullness_tick_threshold", 400.0, 0.0, 1024.0).setTracked(true));
+		HealthyFood.MAX_FULLNESS = register(HealthyFood.MOD_ID + ":generic.max_fullness", new ClampedEntityAttribute("attribute.name.generic.max_fullness", 3.0, 0.0, 1024.0).setTracked(true));
+	}
 }