+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# 1.1.0
+
+- removed dependency on Cloth Config
+- added dependency on Fzzy Config
+- added optional compatibility with Mana Attributes and Stamina Attributes
+- amount of health recovered by eating is now configurable per item
+- eating can now also recover mana and stamina
+
 # 1.0.0
 
 First release.